refactor(upscaling): split PerfMode.cpp into PerfMode/ submodule
Behavior-preserving mechanical split of the 1267-line PerfMode.cpp
monolith into a PerfMode/ sub-module, mirroring the sibling
FoveatedRender/ layout, so the upcoming runtime-quality work has room to
land without growing the monolith further. No logic changes; each member
function's body is relocated verbatim (verified: every PerfMode::
definition migrates exactly once, no duplicates).

Files by concern:
- Core.cpp              SetupResources, DrawSettings (lifecycle)
- ResolutionHook.cpp    BSOpenVR GetRenderTargetSize hook + install
- PostIntercept.cpp     tonemap/refraction/ISCopy IS hooks, downscale, HandlePostProcessing
- MenuBridge.cpp        menu-BG blit, CreateRT enlarge thunks, UI pass
- DSSwap.cpp            fake-DS swap, viewport/fade-overlay/SetDirtyStates hooks
- FullscreenPassScope.cpp  the RAII pipeline-state snapshot helper

The public header stays at Upscaling/PerfMode.h (FoveatedRender precedent),
so no includers change. CMake picks the new files up via the existing
GLOB_RECURSE CONFIGURE_DEPENDS.

Smoke-tested on SkyrimSE: CS + Upscaling load, Streamline/DLSS init, game
reaches in-game gameplay, no errors. PerfMode stays dormant on flat
(VR-only) as expected.
